api_consumer/requirements.txt
kafka-python==1.4.6
psycopg2
sqlalchemy
